Added files via upload
This is the self-contained file including everything useful from files 1,2 and 3. The three "gen" functions create the matrices that are used in *mixture*. Mixture takes as arguments a pre-split text chunk, matrices from the MvS and monte carlo, a BW file from the Baum-Welch, and a set of weights that should sum to 1 to generate text at the length of the original input file. 

The training.mixture function will follow.